Split Dashboard file into smaller files
