+++ src/src/components/Dashboard/index.tsx
@@ -1,22 +1,12 @@
 import React, { CSSProperties } from 'react';
-import { StackedChart } from '../Charts/StackedChart';
-import { GoalsWidget } from '../Widgets/GoalsWidget';
 import { State } from 'typesafe-reducer';
 import { RA } from '../../utils/types';
 import { EventsStore } from '../EventsStore';
-import { DoughnutChart } from '../Charts/DoughnutChart';
-import { Suggestions } from '../Widgets/Suggestions';
-import { QuickActions } from '../Widgets/QuickActions';
-import { DataExport } from '../Widgets/DataExport';
-
-const widgets = {
-  DoughnutChart,
-  StackedChart,
-  DataExport,
-  GoalsWidget,
-  QuickActions,
-  Suggestions,
-} as const;
+import { useBooleanState } from '../../hooks/useBooleanState';
+import { commonText } from '../../localization/common';
+import { useLayout } from './useLayout';
+import { defaultLayout, widgetGridColumnSizes } from './definitions';
+import { Widget } from './Widget';
 
 const breakpointsTailwind = {
   xs: 0,
@@ -43,106 +33,6 @@ export type Widget = {
     | State<'Suggestions'>;
 };
 
-const singleRow = {
-  xs: 1,
-  sm: 1,
-  md: 1,
-  lg: 1,
-  xl: 1,
-  '2xl': 1,
-} as const;
-
-const defaultLayout: RA<Widget> = [
-  {
-    colSpan: {
-      xs: 1,
-      sm: 1,
-      md: 3,
-      lg: 2,
-      xl: 1,
-      '2xl': 1,
-    },
-    rowSpan: singleRow,
-    definition: { type: 'GoalsWidget' },
-  },
-  {
-    colSpan: {
-      xs: 1,
-      sm: 1,
-      md: 3,
-      lg: 2,
-      xl: 1,
-      '2xl': 1,
-    },
-    rowSpan: singleRow,
-    definition: { type: 'QuickActions' },
-  },
-  {
-    colSpan: {
-      xs: 1,
-      sm: 1,
-      md: 3,
-      lg: 2,
-      xl: 1,
-      '2xl': 1,
-    },
-    rowSpan: singleRow,
-    definition: { type: 'Suggestions' },
-  },
-  {
-    colSpan: {
-      xs: 1,
-      sm: 1,
-      md: 3,
-      lg: 3,
-      xl: 3,
-      '2xl': 3,
-    },
-    rowSpan: singleRow,
-    definition: { type: 'StackedChart' },
-  },
-  {
-    colSpan: {
-      xs: 1,
-      sm: 1,
-      md: 2,
-      lg: 2,
-      xl: 2,
-      '2xl': 2,
-    },
-    rowSpan: {
-      xs: 1,
-      sm: 1,
-      md: 2,
-      lg: 2,
-      xl: 2,
-      '2xl': 2,
-    },
-    definition: { type: 'DoughnutChart' },
-  },
-  {
-    colSpan: {
-      xs: 1,
-      sm: 1,
-      md: 3,
-      lg: 2,
-      xl: 1,
-      '2xl': 1,
-    },
-    rowSpan: singleRow,
-    definition: { type: 'DataExport' },
-  },
-];
-
-const widgetGridColumnSizes: WidgetGridColumnSizes = {
-  xs: 1,
-  sm: 1,
-  md: 3,
-  lg: 6,
-  xl: 6,
-  '2xl': 6,
-};
-
 const gridSizeVariables = Object.fromEntries(
   Object.entries(widgetGridColumnSizes).map(([breakpoint, size]) => [
     `--grid-cols-${breakpoint}`,
@@ -155,10 +45,35 @@ export function Dashboard({
 }: {
   readonly durations: EventsStore | undefined;
 }): JSX.Element {
+  const [isEditing, _, __, handleToggle] = useBooleanState();
+
+  const [layout, setLayout] = useLayout();
+  const originalLayout = React.useRef<RA<Widget>>([]);
+  React.useEffect(() => {
+    if (isEditing) originalLayout.current = layout;
+  }, [isEditing]);
+
   return (
-    <div className="overflow-y-auto overflow-x-hidden">
-      <div
-        className={`
+    <div>
+      {isEditing && (
+        <>
+          <button
+            type="button"
+            onClick={(): void => setLayout(originalLayout.current)}
+          >
+            {commonText('cancel')}
+          </button>
+          <button type="button" onClick={(): void => setLayout(defaultLayout)}>
+            {commonText('resetToDefault')}
+          </button>
+        </>
+      )}
+      <button type="button" onClick={handleToggle}>
+        {isEditing ? commonText('save') : commonText('edit')}
+      </button>
+      <div className="overflow-y-auto overflow-x-hidden">
+        <div
+          className={`
           grid grid-cols-[repeat(var(--grid-cols-xs),1fr)] gap-2
           p-2
           sm:grid-cols-[repeat(var(--grid-cols-sm),1fr)]
@@ -167,65 +82,13 @@ export function Dashboard({
           xl:grid-cols-[repeat(var(--grid-cols-xl),1fr)]
           2xl:grid-cols-[repeat(var(--grid-cols-2xl),1fr)]
         `}
-        style={gridSizeVariables as CSSProperties}
-      >
-        {defaultLayout.map((widget, index) => (
-          <Widget widget={widget} key={index} durations={durations} />
-        ))}
+          style={gridSizeVariables as CSSProperties}
+        >
+          {defaultLayout.map((widget, index) => (
+            <Widget widget={widget} key={index} durations={durations} />
+          ))}
+        </div>
       </div>
     </div>
-  );
-}
-
-const widgetClassName = `
-  bg-gray-200 dark:bg-neutral-900 flex flex-col gap-2 rounded p-4
-  col-[span_var(--col-span-xs)_/_span_var(--col-span-xs)]
-  sm:col-[span_var(--col-span-sm)_/_span_var(--col-span-sm)]
-  md:col-[span_var(--col-span-md)_/_span_var(--col-span-md)]
-  lg:col-[span_var(--col-span-lg)_/_span_var(--col-span-lg)]
-  xl:col-[span_var(--col-span-xl)_/_span_var(--col-span-xl)]
-  2xl:col-[span_var(--col-span-2xl)_/_span_var(--col-span-2xl)]
-  row-[span_var(--row-span-xs)_/_span_var(--row-span-xs)] 
-  sm:row-[span_var(--row-span-sm)_/_span_var(--row-span-sm)]
-  md:row-[span_var(--row-span-md)_/_span_var(--row-span-md)]
-  lg:row-[span_var(--row-span-lg)_/_span_var(--row-span-lg)]
-  xl:row-[span_var(--row-span-xl)_/_span_var(--row-span-xl)]
-  2xl:row-[span_var(--row-span-2xl)_/_span_var(--row-span-2xl)]
-`;
-
-function Widget({
-  widget: {
-    colSpan,
-    rowSpan,
-    definition: { type, ...definition },
-  },
-  durations,
-}: {
-  readonly widget: Widget;
-  readonly durations: EventsStore | undefined;
-}): JSX.Element {
-  const colSpanVariables = React.useMemo(
-    () => ({
-      ...Object.fromEntries(
-        Object.entries(colSpan).map(([breakpoint, size]) => [
-          `--col-span${breakpoint === 'xs' ? '' : `-${breakpoint}`}`,
-          size,
-        ])
-      ),
-      ...Object.fromEntries(
-        Object.entries(rowSpan).map(([breakpoint, size]) => [
-          `--row-span${breakpoint === 'xs' ? '' : `-${breakpoint}`}`,
-          size,
-        ])
-      ),
-    }),
-    [colSpan, rowSpan]
-  );
-
-  const WidgetComponent = widgets[type];
-  return (
-    <section style={colSpanVariables} className={widgetClassName}>
-      <WidgetComponent durations={durations} {...definition} />
-    </section>
   );
 }
